requirements.txt
flask>=3.0
flask-cors>=4.0
annotated-types==0.7.0
anthropic==0.92.0
python-dotenv
google-genai
groq
anyio==4.13.0
app-store-scraper==0.3.5
certifi==2026.2.25
chardet==3.0.4
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
google-play-scraper==1.2.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==2.10
jiter==0.13.0
numpy==2.2.6
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.23.0
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==1.25.11
